With GCC, add warning flags "-Wall -W" rather than "-Wall -Wunused" (-Wall implies -Wunused anyway).  Fixed all the warnings this throws up, except in languages/ (that code is to be replaced with Snowball soon).
git-svn-id: file:///home/xapian-svn/xapian/trunk@4233 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/backends/quartz/btree.cc
@@ -716,7 +716,7 @@ void Btree::make_index_item(byte * result, unsigned int result_len,
     }
 
     // FIXME: abort not good - better than buffer overrun though
-    if (I2 + i + C2 + 4 > result_len) abort();
+    if (I2 + i + C2 + 4 > (int)result_len) abort();
 
     SETI(result, 0, I2 + i + C2 + 4); // Set item length
     SETK(result, I2, i + C2);    // Set key length
